pyproject.toml
[tool.poetry]
name = "mallakto-site-backend"
version = "0.1.0"
description = "Django REST framework for Mallakto company web-site"
authors = ["sergey-royt <goodstop687@gmail.com>"]
readme = "README.md"
packages = [{include = "site_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^74.1.3"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.1"
pre-commit = "^3.8.0"
dj-database-url = "^2.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
coverage = "^7.6.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
--line-length = 79
